Try to lock qoiz encoding behind a cargo feature / go build flag.
