+++ web/packages/shared/libs/ironrdp/Cargo.toml
@@ -15,7 +15,6 @@ console_error_panic_hook = "0.1.7"
 getrandom1 = { package = "getrandom", version = "0.2", features = ["js"] }
 getrandom2 = { package = "getrandom", version = "0.4.2", features = ["wasm_js"] }
 ironrdp-session.workspace = true
-ironrdp-session.features = ["qoiz"]
 ironrdp-pdu.workspace = true
 ironrdp-core.workspace = true
 ironrdp-graphics.workspace = true
@@ -28,3 +27,7 @@ tracing-web = "0.1.2"
 uuid = { version = ">=1.16, <1.24", features = ["js"] } # Pinned below 1.21: uuid 1.21+ needs getrandom 0.4 -> rand_core 0.10 (stable), which conflicts with picky rc.22's pinned rand_core = "=0.10.0-rc-3". Remove this pin when picky ships with stable RustCrypto deps.
 wasm-bindgen = "0.2.95"
 web-sys = { version = "0.3.95", features = ["ImageData"] }
+
+[features]
+default = []
+qoiz = ["ironrdp-session/qoiz"]
